[ptr-traits] Add #includes of headers rather than forward declarations for types which are used as pointees in PointerUnions, PointerIntPairs, and DenseMap pointer keys.
This is part of a series of patches to allow LLVM to check for complete
pointee types when computing its pointer traits. This is absolutely
necessary to get correct (or reproducible) results for things like how
many low bits are guaranteed to be zero.

I think this is the last patch for getting Clang clean here!!!