[test-suite] Add sqliterc to silence sqlite3 error
When sqlite3 cannot access the user's home directory (either through
getpwuid or the HOME env var) to find sqliterc, it spits out an error
that causes the result diff to fail. This change adds a blank file and
directs the run options to point sqlite to it to silence the error.